+++ blockchain/state_manager/src/utils.rs
@@ -9,7 +9,8 @@ use forest_actor_interface::{
     power,
 };
 use forest_blocks::Tipset;
-use forest_fil_types::{verifier::ProofVerifier, RegisteredSealProof, SectorInfo, SectorNumber};
+use forest_fil_types::verifier::ProofVerifier;
+use fvm_shared::sector::{RegisteredSealProof, SectorInfo, SectorNumber};
 use forest_ipld_blockstore::BlockStore;
 use fvm_ipld_bitfield::BitField;
 use fvm_shared::address::Address;
