+++ forest/cli/src/cli/fetch_params_cmd.rs
@@ -4,7 +4,7 @@
 use crate::cli::Config;
 use structopt::StructOpt;
 
-use forest_fil_types::SectorSize;
+use fvm_shared::sector::SectorSize;
 use forest_paramfetch::{get_params_default, SectorSizeOpt};
 
 use super::cli_error_and_die;
