+++ node/rpc-api/src/lib.rs
@@ -370,7 +370,8 @@ pub mod state_api {
     use forest_blocks::{
         gossip_block::json::GossipBlockJson as BlockMsgJson, tipset_keys_json::TipsetKeysJson,
     };
-    use forest_fil_types::{deadlines::DeadlineInfo, SectorNumber};
+    use forest_fil_types::deadlines::DeadlineInfo;
+    use fvm_shared::sector::SectorNumber;
     use forest_json::address::json::AddressJson;
     use forest_json::cid::CidJson;
     use forest_message::{message::json::MessageJson, message_receipt::json::MessageReceiptJson};
@@ -522,7 +523,7 @@ pub mod gas_api {
 
 /// Common API
 pub mod common_api {
-    use forest_fil_types::build_version::APIVersion;
+    use super::data_types::APIVersion;
 
     pub const VERSION: &str = "Filecoin.Version";
     pub type VersionParams = ();
