--- types/src/build_version/mod.rs
@@ -1,24 +1,0 @@
-// Copyright 2019-2022 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-use serde::Serialize;
-
-/// Represents the current version of the API.
-#[derive(Serialize)]
-#[serde(rename_all = "PascalCase")]
-pub struct APIVersion {
-    pub version: String,
-    pub api_version: Version,
-    pub block_delay: u64,
-}
-
-/// Integer based value on version information. Highest order bits for Major, Mid order for Minor
-/// and lowest for Patch.
-#[derive(Serialize)]
-pub struct Version(u32);
-
-impl Version {
-    pub const fn new(major: u64, minor: u64, patch: u64) -> Self {
-        Self((major as u32) << 16 | (minor as u32) << 8 | (patch as u32))
-    }
-}
