+++ types/src/lib.rs
@@ -1,12 +1,8 @@
 // Copyright 2019-2022 ChainSafe Systems
 // SPDX-License-Identifier: Apache-2.0, MIT
 
-pub mod build_version;
 pub mod deadlines;
-pub mod sector;
 
 pub mod genesis;
 
 pub mod verifier;
-
-pub use self::sector::*;
