--- types/src/sector/mod.rs
@@ -1,13 +1,0 @@
-// Copyright 2019-2022 ChainSafe Systems
-// SPDX-License-Identifier: Apache-2.0, MIT
-
-pub mod post;
-pub use self::post::*;
-
-pub use fvm_shared::sector::{
-    AggregateSealVerifyInfo, AggregateSealVerifyProofAndInfos, InteractiveSealRandomness,
-    OnChainWindowPoStVerifyInfo, PoStProof, PoStRandomness, RegisteredAggregateProof,
-    RegisteredPoStProof, RegisteredSealProof, SealRandomness, SealVerifyInfo, SealVerifyParams,
-    SectorID, SectorInfo, SectorNumber, SectorQuality, SectorSize, Spacetime, StoragePower,
-    WindowPoStVerifyInfo, WinningPoStVerifyInfo, MAX_SECTOR_NUMBER,
-};
