+++ utils/json/src/lib.rs
@@ -4,3 +4,4 @@ pub mod actor_state;
 pub mod address;
 pub mod bigint;
 pub mod cid;
+pub mod sector;
